+++ api/v1/routing/handlers/handlers.go
@@ -13,7 +13,16 @@ func Pong(w http.ResponseWriter, r *http.Request) {
 
 func ServeIndex(w http.ResponseWriter, r *http.Request) {
 	w.Header().Add("Content-Type", "text/html")
-	content, err := os.ReadFile("./docs/index.html")
+	var file string
+	if r.URL.Path == "/dark" {
+		file = "./docs/dark/index.html"
+	} else if r.URL.Path == "/" || r.URL.Path == "/light" {
+		file = "./docs/index.html"
+	} else {
+		file = "./docs/404/index.html"
+	}
+
+	content, err := os.ReadFile(file)
 
 	if err != nil {
 		response.Error(w, errors.FAILED("serving index.html"))