+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Added get users endpoint
 - Wrote api-spec.yml
 - Added index.c and Makefile
 - User registration